backend/pyproject.toml
[tool.poetry]
name = "comments-spa"
version = "0.1.0"
description = ""
authors = ["AndrewwSS <clash2clans1one@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-recursive = "^0.1.2"
pillow = "^10.4.0"
django-redis = "^5.4.0"
celery = "^5.4.0"
django-cors-headers = "^4.4.0"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"
django-debug-toolbar = "^4.4.6"
pika = "^1.3.2"
django-storages = "^1.14.4"
boto3 = "^1.35.68"
pika-stubs = "^0.1.3"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
